[IE-119] workspace 목록 조회 로딩 상태 UI 수정
